+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This file is similar to the format suggested by [Keep a CHANGELOG](https://githu
 
 ## Unreleased
 - [Feature] Add `hasExternalIcon` to **NavBar** to display an external icon on hover ([#1361](https://github.com/optimizely/oui/pull/1361)).
+- [Feature] New component **SummaryBar** for having key information highlighted in a condensed component ([#1360](https://github.com/optimizely/oui/pull/1360))
 
 ## 46.6.1 - 2020-06-23
 - [Patch] Update **Tile** Prop TypeScript definition for `dropdownItems`. Also updated **Tile** stories to be tsx (for TypeScript validation) and the ability for all Stories to use TypeScript. ([#1359](https://github.com/optimizely/oui/pull/1359))